Unskip API specs from scaffold
This refactors some of the specs generated for the endpoint by the
scaffold. I do not consider these complete but will be merging anyway
since this is not a production application.

To make them complete they will need to test:
* that the error responses are as expected, instead of just "not
  successful".
* Ensure the ShortenedUrl created has the same attributes as those given
  in the params.

There is is authentication and limitations to be added to the Create
screen yet so I wont complete these specs until those are in place.
