--- spec/requests/shortened_urls_spec.rb
@@ -1,78 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-# This spec was generated by rspec-rails when you ran the scaffold generator.
-# It demonstrates how one might use RSpec to test the controller code that
-# was generated by Rails when you ran the scaffold generator.
-#
-# It assumes that the implementation code is generated by the rails scaffold
-# generator. If you are using any extension libraries to generate different
-# controller code, this generated spec may or may not pass.
-#
-# It only uses APIs available in rails and/or rspec-rails. There are a number
-# of tools you can use to make these specs even more expressive, but we're
-# sticking to rails and rspec-rails APIs to keep things simple and stable.
-
-RSpec.describe '/shortened_urls', type: :request do
-  # ShortenedUrl. As you add validations to ShortenedUrl, be sure to
-  # adjust the attributes here as well.
-  let(:valid_attributes) do
-    skip('Add a hash of attributes valid for your model')
-  end
-
-  let(:invalid_attributes) do
-    skip('Add a hash of attributes invalid for your model')
-  end
-
-  describe 'GET /index' do
-    it 'renders a successful response' do
-      ShortenedUrl.create! valid_attributes
-      get shortened_urls_url
-      expect(response).to be_successful
-    end
-  end
-
-  describe 'GET /show' do
-    it 'renders a successful response' do
-      shortened_url = ShortenedUrl.create! valid_attributes
-      get shortened_url_url(shortened_url)
-      expect(response).to be_successful
-    end
-  end
-
-  describe 'GET /new' do
-    it 'renders a successful response' do
-      get new_shortened_url_url
-      expect(response).to be_successful
-    end
-  end
-
-  describe 'POST /create' do
-    context 'with valid parameters' do
-      it 'creates a new ShortenedUrl' do
-        expect do
-          post shortened_urls_url, params: { shortened_url: valid_attributes }
-        end.to change(ShortenedUrl, :count).by(1)
-      end
-
-      it 'redirects to the created shortened_url' do
-        post shortened_urls_url, params: { shortened_url: valid_attributes }
-        expect(response).to redirect_to(shortened_url_url(ShortenedUrl.last))
-      end
-    end
-
-    context 'with invalid parameters' do
-      it 'does not create a new ShortenedUrl' do
-        expect do
-          post shortened_urls_url, params: { shortened_url: invalid_attributes }
-        end.to change(ShortenedUrl, :count).by(0)
-      end
-
-      it "renders a successful response (i.e. to display the 'new' template)" do
-        post shortened_urls_url, params: { shortened_url: invalid_attributes }
-        expect(response).to be_successful
-      end
-    end
-  end
-end
